Refactor: test folder structure
