+++ AuthServiceTests/UnitTests/UserTableQueryTests.cs
@@ -1,20 +1,15 @@
-using AuthService.Authentication;
 using AuthService.DataAccess;
 using AuthService.DataAccess.UserTableQueries;
 using Microsoft.EntityFrameworkCore;
 using AuthServiceModelLibrary.Enums;
 using AuthServiceModelLibrary.ApplicationUser;
-using AuthServiceModelLibrary.DTOs;
 
-namespace AuthServiceTests
+namespace AuthServiceTests.UnitTests
 {
-    [TestFixture]
-    public class AuthOperationsTests
+    public class UserTableQueryTests
     {
         private AuthDbContext _dbContext;
         private IUserTableQueries _queries;
-        private IAuthOperations _operations;
-        private ISecurityUtil _securityUtil;
 
         [SetUp]
         public void Setup()
@@ -24,35 +19,12 @@ namespace AuthServiceTests
                 .UseInMemoryDatabase(databaseName: dbName)
                 .Options;
             _dbContext = new AuthDbContext(options);
+
             _queries = new UserTableQueries(_dbContext);
-            _securityUtil = new SecurityUtil();
-            _operations = new AuthOperations(_queries, _securityUtil);
         }
 
         [Test]
-        public void Register_ValidInput_ReturnsTrue()
-        {
-            // Arrange
-            using (_dbContext)
-            {
-                ApplicationUser user = new ApplicationUser()
-                {
-                    FirstName = "Jolan",
-                    LastName = "Hegyi",
-                    UserName = "Hegyine",
-                    Email = "hegyine@hegy.com",
-                    PasswordHash = "pw",
-                    Salt = "123",
-                    Role = "Administrator"
-                };
-                
-                // Assert
-                Assert.IsTrue(_operations.Register(user));
-            }
-        }
-
-        [Test]
-        public void Register_AddSameUserTwice_ThrowsError()
+        public void AddUser_ValidInput_IsInDb()
         {
             // Arrange
             using (_dbContext)
@@ -68,16 +40,16 @@ namespace AuthServiceTests
                     Role = "Administrator"
                 };
 
-                // Act 
-                _operations.Register(user);
+                // Act
+                _queries.AddUser(user);
 
                 // Assert
-                Assert.Throws<System.ArgumentException>(() => _operations.Register(user));
+                Assert.That(_dbContext.Users.FirstOrDefault(u => u.Email == user.Email), Is.SameAs(user));
             }
         }
 
         [Test]
-        public void Register_AddIncompleteUser_ThrowsError()
+        public void AddUser_AddEmptyModel_ThrowsError()
         {
             // Arrange
             using (_dbContext)
@@ -85,12 +57,12 @@ namespace AuthServiceTests
                 ApplicationUser user = new ApplicationUser();
 
                 // Assert
-                Assert.Throws<DbUpdateException>(() => _operations.Register(user));
+                Assert.Throws<DbUpdateException>(() => _queries.AddUser(user));
             }
         }
 
         [Test]
-        public void Register_ValidInput_IsInDb()
+        public void AddUser_AddSameUserTwice_ThrowsError()
         {
             // Arrange
             using (_dbContext)
@@ -107,99 +79,27 @@ namespace AuthServiceTests
                 };
 
                 // Act
-                _operations.Register(user);
+                _queries.AddUser(user);
 
                 // Assert
-                Assert.That(_dbContext.Users.FirstOrDefault(u => u.UserName == user.UserName), Is.EqualTo(user));
+                Assert.Throws<ArgumentException>(() => _queries.AddUser(user));
             }
         }
 
         [Test]
-        public void Register_AddSameEmailTwice_ThrowsError()
+        public void UpdateUser_ValidInput_IsUpdated()
         {
             // Arrange
             using (_dbContext)
             {
-                ApplicationUser user1 = new ApplicationUser()
-                {
-                    FirstName = "Jolan",
-                    LastName = "Hegyi",
-                    UserName = "Hegyine",
-                    Email = "hegyine@hegy.com",
-                    PasswordHash = "pw",
-                    Salt = "123",
-                    Role = "Administrator"
-                };
-
-                ApplicationUser user2 = new ApplicationUser()
-                {
-                    FirstName = "Jolan2",
-                    LastName = "Hegyi2",
-                    UserName = "Hegyine2",
-                    Email = "hegyine@hegy.com",
-                    PasswordHash = "pw2",
-                    Salt = "1232",
-                    Role = "Administrator"
-                };
-
-                _operations.Register(user1);
-                Assert.Throws<ArgumentException>(() => _operations.Register(user2));
-            }
-        }
-
-        [Test]
-        public void Register_AddSameUsernameTwice_ThrowsError()
-        {
-            // Arrange
-            using (_dbContext)
-            {
-                ApplicationUser user1 = new ApplicationUser()
-                {
-                    FirstName = "Jolan",
-                    LastName = "Hegyi",
-                    UserName = "Hegyine",
-                    Email = "hegyine@hegy.com",
-                    PasswordHash = "pw",
-                    Salt = "123",
-                    Role = "Administrator"
-                };
-
-                ApplicationUser user2 = new ApplicationUser()
-                {
-                    FirstName = "Jolan2",
-                    LastName = "Hegyi2",
-                    UserName = "Hegyine",
-                    Email = "hegyine@hegy.com2",
-                    PasswordHash = "pw2",
-                    Salt = "1232",
-                    Role = "Administrator"
-                };
-
-                // Act
-                _operations.Register(user1);
-
-                // Assert
-                Assert.Throws<ArgumentException>(() => _operations.Register(user2));
-            }
-        }
-
-        [Test]
-        public void Login_ExistingUser_ReturnSameUser()
-        {
-            // Arrange
-            using (_dbContext)
-            {
-                string password = "pw";
-                string salt = _securityUtil.CreateSalt();
-
                 ApplicationUser user = new ApplicationUser()
                 {
                     FirstName = "Jolan",
                     LastName = "Hegyi",
                     UserName = "Hegyine",
                     Email = "hegyine@hegy.com",
-                    Salt = salt,
-                    PasswordHash = _securityUtil.HashPassword(password, salt),
+                    PasswordHash = "pw",
+                    Salt = "123",
                     Role = "Administrator"
                 };
 
@@ -207,56 +107,174 @@ namespace AuthServiceTests
                 _dbContext.Users.Add(user);
                 _dbContext.SaveChanges();
 
-                IUserLoginDTO userDTO = new UserLoginDTO();
-                userDTO.Email = user.Email;
-                userDTO.Password = password;
+                string newEmail = "nemjolan@hegy.com";
+                user.Email = newEmail;
+
+                _queries.UpdateUser(user);
 
                 // Assert
-                Assert.That(user, Is.SameAs(_operations.Login(userDTO)));
+                Assert.That(_dbContext.Users.FirstOrDefault(u => u.Email == user.Email), Is.SameAs(user));
             }
         }
 
         [Test]
-        public void Login_MissingUser_ThrowsError()
-        {
-            // Arrange
-            IUserLoginDTO userLoginDto = new UserLoginDTO();
-            userLoginDto.Email = "NotAnEmail";
-            userLoginDto.Password = "NotAPassword";
-
-            // Assert
-            Assert.Throws<ArgumentException>(() => _operations.Login(userLoginDto));
-        }
-
-        [Test]
-        public void Login_WrongPassword_ThrowsError()
+        public void UpdateUser_MissingUser_ThrowsError()
         {
             // Arrange
             using (_dbContext)
             {
-                string password = "pw";
-                string salt = _securityUtil.CreateSalt();
-
                 ApplicationUser user = new ApplicationUser()
                 {
                     FirstName = "Jolan",
                     LastName = "Hegyi",
                     UserName = "Hegyine",
                     Email = "hegyine@hegy.com",
-                    Salt = salt,
-                    PasswordHash = _securityUtil.HashPassword(password, salt),
+                    PasswordHash = "pw",
+                    Salt = "123",
+                    Role = "Administrator"
+                };
+
+                // Assert 
+                Assert.Throws<DbUpdateConcurrencyException>(() => _queries.UpdateUser(user));
+            }
+        }
+
+        [Test]
+        public void DeleteUser_IsInDb_SuccessfulDelete()
+        {
+            // Arrange
+            using (_dbContext)
+            {
+                ApplicationUser user = new ApplicationUser()
+                {
+                    FirstName = "Jolan",
+                    LastName = "Hegyi",
+                    UserName = "Hegyine",
+                    Email = "hegyine@hegy.com",
+                    PasswordHash = "pw",
+                    Salt = "123",
                     Role = "Administrator"
                 };
 
                 // Act
                 _dbContext.Users.Add(user);
-                IUserLoginDTO userDTO = new UserLoginDTO();
-                userDTO.Email = user.Email;
-                userDTO.Password = "WrongPassword";
+                _dbContext.SaveChanges();
+
+                _queries.DeleteUser(user.Id);
 
                 // Assert
-                Assert.Throws<ArgumentException>(() => _operations.Login(userDTO));
+                Assert.IsNull(_dbContext.Users.FirstOrDefault(u => u.Email == user.Email));
             }
+        }
+
+        [Test]
+        public void DeleteUser_NotInDb_ThrowsError()
+        {
+            // Arrange
+            using (_dbContext)
+            {
+                // Assert
+                Assert.Throws<InvalidOperationException>(() => _queries.DeleteUser("NemId"));
+            }
+        }
+
+        [Test]
+        public void GetUserById_InDb_Successful()
+        {
+            // Arrange
+            using (_dbContext)
+            {
+                ApplicationUser user = new ApplicationUser()
+                {
+                    FirstName = "Jolan",
+                    LastName = "Hegyi",
+                    UserName = "Hegyine",
+                    Email = "hegyine@hegy.com",
+                    PasswordHash = "pw",
+                    Salt = "123",
+                    Role = "Administrator"
+                };
+
+                // Act
+                _dbContext.Users.Add(user);
+                _dbContext.SaveChanges();
+
+                // Assert
+                Assert.That(_queries.GetUserById(user.Id), Is.SameAs(user));
+            }
+        }
+
+        [Test]
+        public void GetUserByEmail_InDb_Successful()
+        {
+            // Arrange
+            using (_dbContext)
+            {
+                ApplicationUser user = new ApplicationUser()
+                {
+                    FirstName = "Jolan",
+                    LastName = "Hegyi",
+                    UserName = "Hegyine",
+                    Email = "hegyine@hegy.com",
+                    PasswordHash = "pw",
+                    Salt = "123",
+                    Role = "Administrator"
+                };
+
+                // Act
+                _dbContext.Users.Add(user);
+                _dbContext.SaveChanges();
+
+                // Assert
+                Assert.That(_queries.GetUserByEmail(user.Email), Is.SameAs(user));
+            }
+        }
+
+        [Test]
+        public void GetUserByUsername_InDb_Successful()
+        {
+            // Arrange
+            using (_dbContext)
+            {
+                ApplicationUser user = new ApplicationUser()
+                {
+                    FirstName = "Jolan",
+                    LastName = "Hegyi",
+                    UserName = "Hegyine",
+                    Email = "hegyine@hegy.com",
+                    PasswordHash = "pw",
+                    Salt = "123",
+                    Role = "Administrator"
+                };
+
+                // Act
+                _dbContext.Users.Add(user);
+                _dbContext.SaveChanges();
+
+                // Assert
+                Assert.That(_queries.GetUserByUsername(user.UserName), Is.SameAs(user));
+            }
+        }
+
+        [Test]
+        public void GetUserById_NotInDb_ReturnsNull()
+        {
+            // Assert
+            Assert.That(_queries.GetUserById("NotAnId"), Is.Null);
+        }
+
+        [Test]
+        public void GetUserByEmail_NotInDb_ReturnsNull()
+        {
+            // Assert
+            Assert.That(_queries.GetUserByEmail("NotAnEmail"), Is.Null);
+        }
+
+        [Test]
+        public void GetUserByUsername_NotInDb_ReturnsNull()
+        {
+            // Assert
+            Assert.That(_queries.GetUserByUsername("NotAUsername"), Is.Null);
         }
     }
 }
